♻️ Refactor: Move data related codes to CoreDataManager
getAllItems(), createItem(), deleteItem(), updateItem()
